Struct splendor_engine::GameContext

pub struct GameContext { /* private fields */ }

A struct to represent the game context.

Implementations

impl GameContext

pub fn random( player_actors: SmallVec<Box<dyn PlayerActor>, MAX_PLAYERS> ) -> Self

Create a new game context with thread_rng.

pub fn with_rng<R: RngCore>( rng: &mut R, player_actors: SmallVec<Box<dyn PlayerActor>, MAX_PLAYERS> ) -> Self

Create a new game context with a given random number generator.

This can be used to create a game context with a specific seed for reproducibility.

pub async fn step( &mut self ) -> Result<Option<SmallVec<usize, MAX_PLAYERS>>, StepError>

Step the game by one turn.

impl GameContext

pub fn n_players(&self) -> usize

Get the number of players in the game.

pub fn last_round(&self) -> bool

Get is the game in the last round.

pub fn game_end(&self) -> bool

Get is the game ended.

pub fn current_round(&self) -> usize

Get the current round.

pub fn current_player(&self) -> usize

Get the current player.

pub fn tokens(&self) -> ColorVec

Get the tokens available in the game.

pub fn snapshot(&self) -> GameSnapshot

Create a snapshot of the game.